Defining, Declaring and Updating Structs in go
We never had to ask a card what its suit or what it value was. We just had the card to be a string that has the suit and value in it like Ace(suit) of Spades(value).

With struct we can have a card with the correct representation of the suit and value. This way we can ask for just what the suit of a card is or the value instead of having to do a string manipulation to get the suit or value of a card from the string.

Structs
Structs means data structure. It is a collection of properties that are related together. Think of struct in go as a Plain Object in Javascript or Java