Struct with Receiver functions & intro to pointers
This comes in handy when you want have your own OOP approach in Go when you may define your have constructors, setters and getters. REMEMBER go is not Object Oriented in nature! You there is an approach to OO with go

Whenever, you create a Struct and assign it to a variable, your computer finds a container or space called address in your RAM and saves that variable. Now your variable will be pointing to that address in your RAM. Any time you print your variable, it will go to the address it points to and prints the value stored in that address

Remember with c++ there are two ways to pass a variable value to a function in C++. Either pass by reference where you pass the original variable (where the real variable is stored in the RAM) or pass by value where you pass a copy of the variable (A copy of your variable is made and stored in another address then your function will only do logic on that cop leaving the original variable and its address intact). Thesame concepts apply in Go