Maps(these as Objects for JavaScript) in Go
Looked at
- How to declare a map
- Intialize a map
- update a map keyValue
- delete keyValue
- Loop through a map
- When to use Maps over Structs

A map is a collection of Key value pairs. Think of a map as Object in JavaScript

When we define maps in go, all the keys must be of thesame type and all the value must be of thesame type as well

Diff between Map and Struct
- With Map all the keys and maps must be of thesame type while with Struct, we have to define the keys ahead of time and the values can be of different type
- With Map, the keys can be indexed meaning, we can loop through the keyValue pairs , while with Struct, we cannot do loop through the keyValues pairs or properties inside struct
- Maps is a reference type while Struct is a value type
-With Maps you don't need to know all the keys or field names ahead of time while while with struct you need to know all your keys or properties ahead of time or at compile time

When to use Map over Structs
- You want to use a map when you are representing an item of very closely related properties
- Use Map when you really don't know how may keys you want before compile time.

In real world, you end up using Structs more often than maps. But again rules when to use then changes depending on your application. With more experience you learn