Interfaces - Problem Interface is trying to solve
Eg if you look at the shuffle function in the Cards Project, the logic in shuffle cards looks like it does not belong to a specific type. How about we want to use that function to shuffle a slice of int or slice of float, we will have to implement thesame logic but pass a different type of receiver. This will be too repetitive. But, With the help of interfaces we can get this done.