Understanding Channels and Go Routines (Async/Await  like JS in Go)
Channels and Go Routines are both structures inside of Go that are used forhandling concurrent programming. You can look at it as way to run asynchronous code in go just the way you will do it in Java or JavaScript

Go Routines are used to create background thread from the main thread to run code. Basically, the when the compiler sees the keyword "go" in front of a function call, create a background thread and run the code

Channels MUST be used communicate between routines. You use channel to make aware to a routine to know when a child routine it created is done executing. If not by default, the parent or main routine will not care and keep executing other code. You declare channels thesame way you declare variables. The data or messages sent between channels must be of thesame type corresponding to however we created the channel.

We basically have a slice of string that contains urls to websites. The we make get request to them. If any of the get request fails we tell the user that one of the sites is down or so.

Blocking call is a function or code that takes sometime to execute

Anytime we are the "go" keyword in front of a function call, we are launching a new go routine.

Your main function runs on the main Routine(called Main thread in Java), as the go Scheduler is running main routine and sees the keyword go in front of a function call, it makes a new go routine(called background thread in Java) to run the code inside that function. As, the compiler is in the background thread, and a Blocking call(like http.Get()) is encountered it that function, due to this function runs a in a go routine (background thread), the go Scheduler will not have to wait, it just returns to the main thread and runs the next line of code.
Anytime the go routine(Background thread) finishes, it can return back the result to the main routine(main thread)

All your go routines are not actually executed truly at thesame time when you have just one CPU and multiple go Routines. Your go routines are executed truly in parallel when you spin up multiple CPU cores to match the number of go routines you created. Overall go scheduler helps manage how the go routines should run where in truly parallel or not truly parallel

Once CPU core
When you have one core you pick only one go routine at a time. Your program or code will run concurrently. This means we have multiple threads executing code. If one thread blocks, another one is picked up and worked on. Scheduler runs one go routine until it finishes or a Blocking call(like http.Get()) is encountered it that function before it returns back to the Main routine to executes the next line of code. So this is not truly parallel.

Multiple CPU core
Running code in Parallel(exact same time) only occurs when you have more than one CPU core. This is when the Scheduler runs your one or more go routines in parallel

Terms learned here
Go Scheduler (listens and know when to switch between routines. It tells the CPU core to basically pick up and routine and run the code)
Main routine(Main thread) where you main code runs
Go routine (Background thread) where you asynchronous code runs
CPU core (you can imagine this as a code executor) Picks up a routine and execute the code in it.

Sending Data to a Channel
Eg: channel <- 5 //send value '5' into this channel
Imagine this a s sending a text message

Receive values out of a channel
Eg myValue <- channel //myValue is a variable to receive the value returned by the channel into. myValue will have to wait (imagine this as await in JavaScript) the for channel to return  a value before it continues in the main thread
Imagine this as receiving a text message from your phone, unlocking your phone and reads the message

Passing a channel value off to functions directly
Eg fmt.Println(<-channel)