Function Literals in GO. Similar to Anonymous function in JavaScript
Problem
We want to continually ping the sites url every 5 seconds

The sleep() function is a blocking code.

For this exercise,
- we want to avoid putting a pause on our main routine (Main thread). When the main routine is paused it cannot receive any other messages through the channel even though the go or child routines (background threads) could be done and send messages back to the main thread through the channel.
- We also want to avoid calling the sleep function in the checkLinkAsync() method as well because as a developer when we call that method, we expects it to check the link immediately and not pause for 5 seconds because doing so

Solution
We want to have this sleep function in  between our Main routine and the checkLinsAsync function. We solve this by having a function Literal run in the background thread where we can now have our sleep function as well as our checkLinkAsync() function call logic. This way we are successfully checking the status of these websites every 5 seconds

Function Literal is an un-named function that we can use to wrap some little chunk of code so that we can execute it at some point in time.